% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_aon.R
\name{assign_aon}
\alias{assign_aon}
\title{AON assignment}
\usage{
assign_aon(graph, from, to, flows)
}
\arguments{
\item{graph}{graph object created with configure_graph() function}

\item{from}{Vector or matrix of points from which aggregate flows are to be calculated}

\item{to}{Vector or matrix of points to which aggregate flows are to be calculated}

\item{flows}{Matrix of flows with nrow(flows)==length(from) and ncol(flows)==length(to)}
}
\value{
graph object with assignment on 'flow' column
}
\description{
Perform the all-or-nothing assignment of
Frank-Wolfe algorithm in the User Equilibrium
Traffic Assignment Model. This function is a wrapper for dodgr 'R/dodgr_flows_aggregate.R'
for more information on the function see dodgr documentation
}
