% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_demand.R
\name{configure_demand}
\alias{configure_demand}
\title{Demand configuration}
\usage{
configure_demand(demand, zones)
}
\arguments{
\item{demand}{demand_file source file with demand data}

\item{zones}{zones object created with configure_graph()}
}
\value{
demand object compatible with solve_ue() function
}
\description{
Tranforms demand data from source origin and destination to match origin and destination from graph/zone objects
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones)
graph <- flist[[1]]
zones <- flist[[2]]
demand <- configure_demand(sioux_demand,zones)

}
