% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_graph.R
\name{configure_graph}
\alias{configure_graph}
\title{Graph configuration}
\usage{
configure_graph(
  network,
  zones,
  turn_penalty = F,
  left_side = F,
  addvalue_cols = F,
  weighted = F,
  performance_measures = F,
  use_cost_col = F
)
}
\arguments{
\item{network}{sf object containing network data}

\item{zones}{sf object containing zone data}

\item{turn_penalty}{Adds turn penalty (see dodgr documentation on turn penalty from function 'R/weight-streetnet.R')}

\item{left_side}{Does traffic travel on the left side of the road (TRUE) or the right side (FALSE)? - only has effect on turn angle calculations for edge times. (see dodgr documentation on turn penalty from function 'R/weight-streetnet.R')}

\item{addvalue_cols}{Does network shapefile has columns 'addv_fw' and 'addv_bw' for added cost in links? (TRUE) or (FALSE)}

\item{weighted}{Does network shapefile has columns 'wg_fw' and 'wg_bw' for a cost multiplication factor in links? (TRUE) or (FALSE)}

\item{performance_measures}{Does network shapefile has columns 'alpha' and 'beta' for usage on BPR funcion? (TRUE) or (FALSE)}

\item{use_cost_col}{Use time column as cost (opposite to speed and distance)? (TRUE) or (FALSE)}
}
\value{
graph object ready for usage on solve_ue() function
}
\description{
Tranforms network data from source network and zones
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones,use_cost_col=TRUE)
graph <- flist[[1]]
zones <- flist[[2]]

}
