% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_geography}
\alias{error_geography}
\title{Error geography}
\usage{
error_geography(x)
}
\arguments{
\item{x}{The type of geography in the get_* function.}
}
\description{
Error geography
}
\keyword{internal}
