% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map events FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number indicating the state can be any number from 1 through 56.}

\item{year}{the year for the accident events}
}
\value{
a plot of the state with points for each accident
}
\description{
Reads in data for a particular year and maps the events in a given state specified by state number
}
\note{
Warning: This function appears to be slightly buggy not returning all events on first plotting
}
\examples{
#Plot California accident patterns
fars_map_state(state.num=6, year=2013)
}
