% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data set file from CSV}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{file path as character string.}
}
\value{
a tbl_df of the selected file if is csv if it exists
}
\description{
Reads in a CSV from a file location as tbl_df data for FARS.
}
\note{
Warning: Function will not work if not given valid file name.
}
\examples{
fars_read("accident_2013.csv.bz2")
}
