% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads in FARS Data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{string or numeric that indicates the year of the FARS data that the file name will be created to read.}
}
\value{
list of tbl_dfs with month and year columns
}
\description{
Reads in a list of years and extracts the year and month for each row from the files for each year associated
}
\note{
Warning: Will not work if years are not 2013, 2014, or 2015 as those are the only years with data files
 @importFrom dplyr select mutate
}
\examples{
make_filename(2013)
}
