% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarises FARS data for multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{string or numeric that indicates the years to be summarised.}
}
\value{
tbl_df with columns being the years summarised and rows being the months and each entry being the number of events
}
\description{
Reads in data for each year and summarises the counts of events by month and year.
}
\note{
Warning: Will not work if years are not in 2013, 2014, or 2015 as those are the only years with data files
}
\examples{
\dontrun{fars_summarize_years(2013:2014)}
}
