% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_traffic_region_historic}
\alias{fetch_traffic_region_historic}
\title{Fetch historic regional road traffic data}
\usage{
fetch_traffic_region_historic()
}
\value{
A tibble showing the annual total volume of road traffic across
each GB region since 1993.
}
\description{
\code {fetch_traffic_region_historic} fetches the annual total volume of road
traffic across each GB region since 1993. Results are returned as a tibble.
}
\examples{
fetch_traffic_region_historic()

}
