% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_traffic_road}
\alias{fetch_traffic_road}
\title{Fetch road traffic data for a road}
\usage{
fetch_traffic_road(road = NA, year = NA, direction = FALSE,
  raw = FALSE)
}
\arguments{
\item{year}{A numeric no smaller than 2000.}

\item{direction}{A boolean which if TRUE shows the average annual daily flow
of road traffic by direction of travel. It is set FALSE by default.}

\item{raw}{A boolean which if TRUE shows the actual manual counts taken at
count point locations. Raw counts are collected by trained enumerators and
are used to feed into calculating the average daily flow estimates.
It is set to FALSE by default.}

\item{region}{A character string of the official road name
of interest written within a pair of single or double qoutes.}
}
\value{
A tibble showing the average annual daily flow of traffic for
a given road.
}
\description{
\code {fetch_traffic_road} fetches the average annual daily flow
of road raffic recorded on all count points along a road
since the year 2000. Results are returned as a tibble.
}
\examples{
fetch_traffic_road(road = "A45")
fetch_traffic_road(road = c("A45", "A665"), year = 2000)
fetch_traffic_road(road = "A45", year = 2000:2005)
fetch_traffic_road(road = "A45", year = c(2000, 2010, 2018))
fetch_traffic_road(road = "A45", direction = TRUE)
fetch_traffic_road(road = "A45", year = 2000, raw = TRUE)

}
