% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get paths}
\usage{
get_paths(graph, demand)
}
\arguments{
\item{graph}{Graph object created with configure_graph() function}

\item{demand}{Demand object created with configure_demand() function}
}
\value{
List of list of paths tracing all connections between nodes such that if x <- dodgr_paths (graph, from, to), then the path between from[i] and to[j] is x [[i]] [[j]].
}
\description{
Wrapper function for dodgr function 'R/dodgr_paths.R'
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones,use_cost_col=TRUE)
graph <- flist[[1]]
zones <- flist[[2]]
demand <- configure_demand(sioux_demand,zones)
paths <- get_paths(graph,demand)

}
