% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\title{Get times}
\usage{
get_times(graph, demand)
}
\arguments{
\item{graph}{Graph object created with configure_graph() function}

\item{demand}{Demand object created with configure_demand() function}
}
\value{
Square matrix of distances between nodes
}
\description{
Wrapper function for dodgr function 'R/dodgr_times.R'
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones,use_cost_col=TRUE)
graph <- flist[[1]]
zones <- flist[[2]]
demand <- configure_demand(sioux_demand,zones)
paths <- get_times(graph,demand)

}
