% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_traffic_all}
\alias{get_traffic_all}
\title{Get all road traffic data}
\usage{
get_traffic_all(direction = FALSE)
}
\arguments{
\item{direction}{A boolean which if TRUE shows the average annual daily flow
of road traffic by direction of travel. It is set FALSE by default.}
}
\value{
A tibble showing the average annual daily flow of traffic for
all counts in Great Britain.
}
\description{
\code {get_traffic_all} gets the average annual daily flow
of road raffic recorded on all count points within Great Britain since
the year 2000. Results are returned as a tibble.
}
\examples{
get_traffic_all()
get_traffic_all(direction = TRUE)
}
