% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_traffic_authority}
\alias{get_traffic_authority}
\title{Get road traffic data for a local authority}
\usage{
get_traffic_authority(authority = NA, year = NA, direction = FALSE,
  raw = FALSE)
}
\arguments{
\item{authority}{A character string of the GB local authority
of interest written within a pair of single or double qoutes.}

\item{year}{A numeric no smaller than 2000.}

\item{direction}{A boolean which if TRUE shows the average annual daily flow
of road traffic by direction of travel. It is set FALSE by default.}

\item{raw}{A boolean which if TRUE shows the actual manual counts taken at
count point locations. Raw counts are collected by trained enumerators and
are used to feed into calculating the average daily flow estimates.
It is set to FALSE by default.}
}
\value{
A tibble showing the average annual daily flow of traffic for
a given local authority.
}
\description{
\code {get_traffic_authority} gets the average annual daily flow
of road raffic recorded on all count points within a given GB local
authority since the year 2000. Results are returned as a tibble.
}
\examples{
get_traffic_authority(authority = "Swansea")
get_traffic_authority(authority = c("Swansea", "Lambeth"), year = 2000)
get_traffic_authority(authority = "Swansea", year = 2000:2005)
get_traffic_authority(authority = "Swansea", year = c(2000, 2010, 2018))
get_traffic_authority(authority = "Swansea", direction = TRUE)
get_traffic_authority(authority = "Swansea", year = 2000, raw = TRUE)
}
