% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_traffic_authority_historic}
\alias{get_traffic_authority_historic}
\title{Get historic authority road traffic data}
\usage{
get_traffic_authority_historic()
}
\value{
A tibble showing the annual total volume of road traffic across
each GB local authority since 1993.
}
\description{
\code {get_traffic_authority_historic} gets the annual total volume of road
traffic across each GB local authority since 1993.
Results are returned as a tibble.
}
\examples{
get_traffic_authority_historic()
}
