% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_section.R
\name{golden_section}
\alias{golden_section}
\title{FW Equilibrium assignment}
\usage{
golden_section(flow, auxiliary_flow, graph, accuracy = 1e-08)
}
\arguments{
\item{flow}{Past iteration calculated traffic flow}

\item{auxiliary_flow}{Current iteration calculated traffic flow}

\item{graph}{Graph object created with configure_graph() function}

\item{accuracy}{Golden section accuracy value. The accuracy is suggested to be set as 1e-8.}
}
\value{
optima theta from golden section technique
}
\description{
The golden-section search is a technique for
finding the extremum of a strictly unimodal
function by successively narrowing the range
of values inside which the extremum is known
to exist. For more details please refer to:
https://en.wikipedia.org/wiki/Golden-section_search
Initial params, notice that in our case the
optimal theta must be in the interval [0, 1]
}
