% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_to_sf.R
\name{graph_to_sf}
\alias{graph_to_sf}
\title{Outputs traffic loaded network file}
\usage{
graph_to_sf(graph, network)
}
\arguments{
\item{graph}{graph object created with configure_graph() function}

\item{network}{sf object containing network data}
}
\value{
Simple feature collection containing network data
}
\description{
Merges information from graph object into source network file, generating a shapefile with traffic flow information
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones,use_cost_col=TRUE)
graph <- flist[[1]]
zones <- flist[[2]]
demand <- configure_demand(sioux_demand,zones)
graph <- solve_ue(graph,demand,col = 'demand')
network <- graph_to_sf(graph,sioux_network)

}
