% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_convergent.R
\name{is_convergent}
\alias{is_convergent}
\title{Convergency test}
\usage{
is_convergent(flow, new_flow, accuracy = 1e-06)
}
\arguments{
\item{flow}{Past iteration calculated traffic flow}

\item{new_flow}{Current iteration calculated traffic flow}

\item{accuracy}{Error accuracy value. The accuracy is suggested to be set as 1e-6.}
}
\value{
Boolen (TRUE) or (FALSE)
}
\description{
Regard those two link flows lists as the point
in Euclidean space R^n, then judge the convergence
under given accuracy criterion.
Here the formula
ERR = || x_{k+1} - x_{k} || / || x_{k} ||
is recommended.
}
