% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_time_performance.R
\name{link_time_performance}
\alias{link_time_performance}
\title{Link time performance}
\usage{
link_time_performance(flow, t0, capacity, alpha = 0.15, beta = 4)
}
\arguments{
\item{flow}{Vector with current traffic flow}

\item{t0}{Link free-flow times}

\item{capacity}{Link capacity}

\item{alpha}{alpha BPR function parameter}

\item{beta}{beta BPR function parameter}
}
\value{
Vector with link times adjusted by link flow
}
\description{
Performance function, which indicates the relationship
between flows (traffic volume) and travel time on
the same link. According to the suggestion from Federal
Highway Administration (FHWA) of America, we could use
the following function:
t = t0 * (1 + alpha * (flow / capacity))^beta
}
