% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_time_performance_integrated.R
\name{link_time_performance_integrated}
\alias{link_time_performance_integrated}
\title{Integration of link time performance}
\usage{
link_time_performance_integrated(flow, t0, capacity, alpha = 0.15, beta = 4)
}
\arguments{
\item{flow}{Vector with current traffic flow}

\item{t0}{Link free-flow times}

\item{capacity}{Link capacity}

\item{alpha}{alpha BPR function parameter}

\item{beta}{beta BPR function parameter}
}
\value{
Vector with integrated values from performance function
}
\description{
The integrated (with repsect to link flow) form of
aforementioned performance function.
Some optimization should be implemented for avoiding overflow
}
