% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create File Name FARS Dataset}
\usage{
make_filename(year)
}
\arguments{
\item{year}{string or numeric that indicates the year of the FARS data that the file name will be created to read.}
}
\value{
a string that is the file name to be used in fars_read for the input year
}
\description{
Takes either a number or a string that is the year with no spaces to create a file name for a FARS CSV to read in the data
}
\note{
Warning: Will not work if year is not 2013, 2014, or 2015 as those are the only years with data files
}
\examples{
 \dontrun{make_filename(2013)}
}
