% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_function.R
\name{object_function}
\alias{object_function}
\title{Objective function}
\usage{
object_function(mixed_flow, graph)
}
\arguments{
\item{mixed_flow}{Vector with flows used by golden section search technique}

\item{graph}{Graph object created with configure_graph() function}
}
\value{
Vector with objective function calculation from flow
}
\description{
Objective function in the linear search step
of the optimization model of user equilibrium
traffic assignment problem, the only variable
is mixed_flow in this case.
}
