% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowmap.R
\name{plot_flowmap}
\alias{plot_flowmap}
\title{Plot flowmap}
\usage{
plot_flowmap(
  network,
  flow_factor = 1,
  scale_factor = 10,
  show_labels = FALSE,
  bbox
)
}
\arguments{
\item{network}{Simple feature collection containing network with flow data}

\item{flow_factor}{Exponential factor for flow map display}

\item{scale_factor}{Tmap scale factor for tm_lines function}

\item{show_labels}{Should map (tmap) display text labels? (TRUE) or (FALSE)}

\item{bbox}{sf bbox object}
}
\value{
Nothing; the function outputs in plot window
}
\description{
Quick timap plot from flows in network
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones,use_cost_col=TRUE)
graph <- flist[[1]]
zones <- flist[[2]]
demand <- configure_demand(sioux_demand,zones)
graph <- solve_ue(graph,demand,col = 'demand')
network <- graph_to_sf(graph,sioux_network)
plot_flowmap(network,show_labels = TRUE)

}
