% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_columns_download}
\alias{process_columns_download}
\title{Process columns of download result}
\usage{
process_columns_download(x)
}
\description{
Process columns of download result
}
\keyword{internal}
