% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geography.R
\name{process_constituency}
\alias{process_constituency}
\title{Process constituency name}
\usage{
process_constituency(constituency = NA)
}
\description{
Process constituency name
}
\keyword{internal}
