% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_url.R
\name{process_direction_raw}
\alias{process_direction_raw}
\title{Process count point direction}
\usage{
process_direction_raw(direction, raw)
}
\description{
Process count point direction
}
\keyword{internal}
