% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{process_error}
\alias{process_error}
\title{Process potential error messages}
\usage{
process_error(geography, column_name, message, year, direction, raw)
}
\arguments{
\item{column_name}{The column name in the hidden dataframe corresponding
to the geography argument.}

\item{message}{Message to be displayed in errors.}

\item{year}{The 'year' arguement from the get_* function.}

\item{direction}{The 'direction' arugment from the get_* function.}

\item{raw}{The 'raw' argument from the get_* function.}

\item{geogrpahy}{The geography of the get_* function.}
}
\description{
Process potential error messages
}
\keyword{internal}
