% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_url.R
\name{process_filters}
\alias{process_filters}
\title{Process API filters}
\usage{
process_filters(
  count_point_id = NA,
  region_id = NA,
  local_authority_id = NA,
  year = NA,
  count_method = NA,
  road_name = NA,
  road_category = NA
)
}
\description{
Process API filters
}
\keyword{internal}
