% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_url.R
\name{process_max_page}
\alias{process_max_page}
\title{Process maximum page number of pagination}
\usage{
process_max_page(api_url)
}
\description{
Process maximum page number of pagination
}
\keyword{internal}
