% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sioux_demand}
\alias{sioux_demand}
\title{Demand data for Sioux Falls network}
\format{A data frame with 528 rows and 3 variables:
\describe{
  \item{orig}{id of origin}
  \item{dest}{id of destination}
  \item{demand}{number of vehicles in respect to the OD pair}
  ...
}}
\usage{
sioux_demand
}
\description{
A dataset containing an OD matrix of vehicles for the Sioux Falls network
}
\keyword{datasets}
