% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sioux_network}
\alias{sioux_network}
\title{Network data for Sioux Falls network}
\format{A data frame with 38 rows and 13 variables:
\describe{
  \item{id_link}{link id (integer)}
  \item{ffs_fw}{forward free flow speed in km/h (numeric)}
  \item{ffs_bw}{backward free flow speed in km/h (numeric)}
  \item{cap_fw}{forward link capacity in veh/h (numeric)}
  \item{cap_bw}{backward link capacity in veh/h (numeric)}
  \item{addv_fw}{forward additional cost (numeric)}
  \item{addv_bw}{backward additional cost (numeric)}
  \item{wg_fw}{forward cost weight (numeric)}
  \item{wg_bw}{backward cost weight (numeric)}
  \item{alpha}{BPR alpha parameter (numeric)}
  \item{beta}{BPR beta parameter (numeric)}
  \item{cost}{actual cost (numeric)}
  \item{geometry}{sf object geometry}
  ...
}}
\usage{
sioux_network
}
\description{
A sf object containing the road network
}
\keyword{datasets}
