% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sioux_zones}
\alias{sioux_zones}
\title{Zone data for Sioux Falls network}
\format{A data frame with 528 rows and 3 variables:
\describe{
  \item{id_zone}{zone id (integer)}
  \item{x}{Longitude in WGS84 (numeric)}
  \item{y}{Latitude in WGS84 (numeric)}
  \item{geometry}{sf object geometry}
  ...
}}
\usage{
sioux_zones
}
\description{
A sf object containing the zones to be used with the network
}
\keyword{datasets}
