% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ue.R
\name{solve_ue}
\alias{solve_ue}
\title{FW Equilibrium assignment}
\usage{
solve_ue(graph, demand, col, max_iterations = 200)
}
\arguments{
\item{graph}{Graph object created with configure_graph() function}

\item{demand}{Demand object created with configure_demand() function}

\item{col}{Column from demand file selected for assignment}

\item{max_iterations}{Maximum number of iterations for Frank-Wolfe algorithm}
}
\value{
Graph object with assignment on 'flow' column
}
\description{
Solve the traffic flow assignment model (user equilibrium) by Frank-Wolfe algorithm (all the necessary data must be  properly input into the model in advance).
}
\examples{
flist <- configure_graph(sioux_network,sioux_zones,use_cost_col=TRUE)
graph <- flist[[1]]
zones <- flist[[2]]
demand <- configure_demand(sioux_demand,zones)
graph <- solve_ue(graph,demand,col = 'demand')

}
